% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_ops.R
\name{api_req}
\alias{api_req}
\title{Basic Aquarius API request}
\usage{
api_req(url, username, password)
}
\arguments{
\item{url}{the api endpoint url of interest, character type}

\item{username}{API username required}

\item{password}{API password required}
}
\value{
a BC gov aquarius specific api request
}
\description{
Basic Aquarius API request
}
\examples{
req <- api_req("ts_url", "username", "password")
}
