% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_visits.R
\name{extract_fv_info}
\alias{extract_fv_info}
\title{Retrieves field visit dataset that provides information on length of time for user upload}
\usage{
extract_fv_info(site, username, password)
}
\arguments{
\item{site}{The site name for retrieving field visit information}

\item{username}{API username required}

\item{password}{API password required}
}
\value{
A dataframe that provides the number of days from field visit to upload
and number of site visits conducted each year.
}
\description{
Retrieves field visit dataset that provides information on length of time for user upload
}
\examples{
field_data <- extract_fv_info("08MH0041", "username", "password")

}
