% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{retrieve_approvals}
\alias{retrieve_approvals}
\title{Retrieve the approval status and the change in approvals over time by timeseries}
\usage{
retrieve_approvals(sample_ts, username, password)
}
\arguments{
\item{sample_ts}{The unique ID of a timeseries (several within each site)}

\item{username}{API username required}

\item{password}{API password required}
}
\value{
A dataframe with approval levels and date of approval level change by unique timeseries id
}
\description{
Retrieve the approval status and the change in approvals over time by timeseries
}
\examples{
extract_appr <- retrieve_approvals(
  sample_ts = "8a0ed03b7fad4a40904122d3b6d0086f",
  "username", "password"
)

}
